backend/whatsapp-bot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
supabase==2.10.0
httpx==0.27.0
requests==2.32.0
beautifulsoup4==4.12.3
Pillow==10.4.0
pyzbar==0.1.9
openai==1.50.0
pydantic==2.9.0
pydantic-settings==2.5.2
apscheduler==3.10.4
